+++ lib/data/repository/kakao_auth_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import '../../core/social_login/social_login.dart';
-import '../datasources/kakao_auth_remote_data_source.dart';
-import '../../domain/repositories/kakao_auth_repository.dart';
+import '../data_source/kakao_auth_remote_data_source.dart';
+import '../../domain/repository/kakao_auth_repository.dart';
 
 class KakaoAuthRepositoryImpl implements KakaoAuthRepository {
   final SocialLogin socialLogin;
