+++ lib/domain/usecase/login_with_kakao.dart
@@ -1,5 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import '../../domain/repositories/kakao_auth_repository.dart';
+import '../repository/kakao_auth_repository.dart';
 
 class LoginWithKakao {
   final KakaoAuthRepository repository;
