+++ lib/main.dart
@@ -10,15 +10,14 @@ import 'package:mercenaryhub/core/shared_prefs/shared_prefs.dart';
 import 'package:mercenaryhub/firebase_options.dart';
 import 'package:mercenaryhub/presentation/pages/home/home_page.dart';
 import 'package:mercenaryhub/presentation/pages/mercenary_applicants/mercenary_applicants_page.dart';
-import 'package:mercenaryhub/presentation/pages/splash/splash_view.dart';
+import 'package:mercenaryhub/presentation/pages/splash/splash_page.dart';
 import 'package:mercenaryhub/presentation/pages/setting/setting_page.dart';
 import 'package:mercenaryhub/presentation/pages/setting/alarm_setting_page.dart';
 import 'package:mercenaryhub/presentation/pages/team_apply_history/team_apply_history_page.dart';
 import 'package:mercenaryhub/presentation/pages/mercenary_apply_history/mercenary_apply_history_page.dart';
 import 'package:mercenaryhub/presentation/pages/login/login_view.dart';
 import 'package:mercenaryhub/presentation/pages/team_invitation_history/team_invitation_history_page.dart';
-import 'package:mercenaryhub/presentation/pages/terms/widget/terms_page.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:mercenaryhub/presentation/pages/terms/terms_page.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -41,7 +40,7 @@ class MainApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: ThemeData(fontFamily: 'Pretendard'),
-      home: const SplashView(),
+      home: const SplashPage(),
       routes: {
         '/setting': (context) => const SettingPage(),
         '/alarm_setting': (context) => const AlarmSettingPage(),
