+++ lib/presentation/pages/home/widgets/home_app_bar.dart
@@ -6,7 +6,7 @@ import 'package:mercenaryhub/presentation/pages/home/view_models/mercenary_feed_
 import 'package:mercenaryhub/presentation/pages/home/view_models/team_feed_view_model.dart';
 import 'package:mercenaryhub/presentation/pages/home/widgets/feed_type_dialog.dart';
 import 'package:mercenaryhub/presentation/pages/write.dart/write_page.dart';
-import 'package:mercenaryhub/presentation/pages/write.dart/write_view_model.dart';
+import 'package:mercenaryhub/presentation/pages/write.dart/view_models/write_view_model.dart';
 
 class HomeAppBar extends StatefulWidget implements PreferredSizeWidget {
   @override
