+++ lib/presentation/pages/terms/terms_page.dart
