+++ lib/presentation/pages/write.dart/view_models/write_view_model.dart
@@ -2,67 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:intl/intl.dart';
-import 'package:mercenaryhub/domain/entity/time_state.dart';
-import 'package:mercenaryhub/presentation/pages/providers.dart';
-
-class WriteState {
-  bool isErrorVisible;
-  bool isLocationFieldEnable;
-  bool isDateFieldEnable;
-  bool isStartTimeFieldEnable;
-  bool isEndTimeFieldEnable;
-  XFile? imageFile;
-  String? imageUrl;
-  String? location;
-  String? level;
-  DateTime date;
-  TimeState time;
-
-  WriteState({
-    required this.isErrorVisible,
-    required this.isLocationFieldEnable,
-    required this.isDateFieldEnable,
-    required this.isStartTimeFieldEnable,
-    required this.isEndTimeFieldEnable,
-    required this.imageFile,
-    required this.imageUrl,
-    required this.location,
-    required this.level,
-    DateTime? date,
-    TimeState? time,
-  })  : date = date ?? DateTime.now(),
-        time = time ?? TimeState();
-
-  WriteState copyWith({
-    bool? isErrorVisible,
-    bool? isLocationFieldEnable,
-    bool? isDateFieldEnable,
-    bool? isStartTimeFieldEnable,
-    bool? isEndTimeFieldEnable,
-    XFile? imageFile,
-    String? imageUrl,
-    String? location,
-    String? level,
-    DateTime? date,
-    TimeState? time,
-  }) {
-    return WriteState(
-      isErrorVisible: isErrorVisible ?? this.isErrorVisible,
-      isLocationFieldEnable:
-          isLocationFieldEnable ?? this.isLocationFieldEnable,
-      isDateFieldEnable: isDateFieldEnable ?? this.isDateFieldEnable,
-      isStartTimeFieldEnable:
-          isStartTimeFieldEnable ?? this.isStartTimeFieldEnable,
-      isEndTimeFieldEnable: isEndTimeFieldEnable ?? this.isEndTimeFieldEnable,
-      imageFile: imageFile ?? this.imageFile,
-      imageUrl: imageUrl ?? this.imageUrl,
-      location: location ?? this.location,
-      level: level ?? this.level,
-      date: date ?? this.date,
-      time: time ?? this.time,
-    );
-  }
-}
+import 'package:mercenaryhub/providers/feed_providers.dart';
+import 'package:mercenaryhub/presentation/pages/write.dart/models/write_state.dart';
 
 class WriteViewModel extends AutoDisposeNotifier<WriteState> {
   @override
