+++ lib/presentation/pages/write.dart/widgets/image_bottom_sheet.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:mercenaryhub/core/loading_bar.dart';
-import 'package:mercenaryhub/presentation/pages/write.dart/write_view_model.dart';
+import 'package:mercenaryhub/presentation/pages/write.dart/view_models/write_view_model.dart';
 
 class ImageBottomSheet extends StatelessWidget {
   final EdgeInsets scaffoldPadding;
