+++ lib/presentation/pages/write.dart/widgets/level_dropdown_form_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mercenaryhub/presentation/pages/write.dart/write_view_model.dart';
+import 'package:mercenaryhub/presentation/pages/write.dart/view_models/write_view_model.dart';
 
 class LevelDropdownFormField extends StatelessWidget {
   final String typeText;
