+++ lib/providers/kakao_login_providers.dart
@@ -2,10 +2,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:firebase_auth/firebase_auth.dart' as firebase_auth;
 import '../core/social_login/kakao_social_login.dart';
 import '../core/social_login/social_login.dart';
-import '../data/datasources/kakao_auth_remote_data_source.dart';
-import '../data/repositories/kakao_auth_repository_impl.dart';
-import '../domain/repositories/kakao_auth_repository.dart';
-import '../domain/usecases/login_with_kakao.dart';
+import '../data/data_source/kakao_auth_remote_data_source.dart';
+import '../data/repository/kakao_auth_repository_impl.dart';
+import '../domain/repository/kakao_auth_repository.dart';
+import '../domain/usecase/login_with_kakao.dart';
 import '../presentation/pages/login/view_models/kakao_login_viewmodel.dart';
 
 final kakaoLoginProvider = Provider<SocialLogin>((ref) => KakaoLogin());
